+++ sdks/workflow-browser-sdk/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ballerine/workflow-browser-sdk
 
+## 0.5.11-3e08f108.6
+
+### Patch Changes
+
+- Pre bump
+- Updated dependencies
+  - @ballerine/workflow-core@0.5.11-3e08f108.6
+  - @ballerine/common@0.7.13-3e08f108.5
+
+## 0.5.11-3e08f108.5
+
+### Patch Changes
+
+- bump pre
+- Updated dependencies
+  - @ballerine/workflow-core@0.5.11-3e08f108.5
+  - @ballerine/common@0.7.13-3e08f108.4
+
 ## 0.5.11-3e08f108.4
 
 ### Patch Changes