+++ services/websocket-service/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ballerine/websocket-service
 
+## 0.0.9-3e08f108.5
+
+### Patch Changes
+
+- Pre bump
+
+## 0.0.9-3e08f108.4
+
+### Patch Changes
+
+- bump pre
+
 ## 0.0.9-3e08f108.3
 
 ### Patch Changes